requirements.txt
torch==2.0.1
torchvision==0.15.2
apache-airflow==2.7.3
dvc==3.43.1
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
opencv-python==4.8.0.76
Pillow==10.0.0
tqdm==4.65.0
python-dotenv==1.0.0